+++ source/libs/sgeobj/sge_answer.c
@@ -361,7 +361,7 @@ void answer_to_dstring(const lListElem *answer, dstring *diag)
          const char *s, *t;
          s = lGetString(answer, AN_text);
          if ((t=strchr(s, '\n'))) {
-            sge_dstring_sprintf_append(diag, "%.*s", t-s, s); 
+            sge_dstring_sprintf_append(diag, "%.*s", (int) (t-s), s);
          }
          else {
             sge_dstring_append(diag, s);
